finally got GrandFinal working
